This includes Marick's addition of names for forms. I changed this to provide both 'name' and 'id'. I'm still pretty confused about this and am not quite sure whether we need both or note.
I've also updated show_elements to show these names (and also the
index). show_elements is now also a method of forms.

One thing that i've noticed is that if you do forms('start'), you get
a collection object containing two forms, because there are two forms
with that name. This a potential source of confusion. (Heretofore, the
functions we provided provide the first if there was more than one.)

All of the doc and examples still need to be updated for this stuff. I
still need to play with it a bit.