Fixes a performance bug in Watir.
http://jira.openqa.org/browse/WTR-193
Previously, we had optimized Watir to find elements by name/id. Some time ago, this optimization broke, and Watir reverted to using the old-fashioned method, which was slow.

It now works, and what is more, there is now a unit test that will detect this kind of problem in the future.